Fix bug with devices

+++ neural_clbf/experiments/common/episodic_datamodule.py
@@ -110,7 +110,7 @@ class EpisodicDataModule(pl.LightningDataModule):
         x_sim = self.sample_trajectories(self.model.nominal_simulator)
 
         # Augment those points with samples from the fixed range
-        x_sample = self.sample_fixed().type_as(x_sim)
+        x_sample = self.sample_fixed()
         x = torch.cat((x_sim, x_sample), dim=0)
 
         # Randomly split data into training and test sets
@@ -148,7 +148,7 @@ class EpisodicDataModule(pl.LightningDataModule):
         x_sim = self.sample_trajectories(simulator)
 
         # Augment those points with samples from the fixed range
-        x_sample = self.sample_fixed().type_as(x_sim)
+        x_sample = self.sample_fixed()
         x = torch.cat((x_sim, x_sample), dim=0)
 
         print(f"Sampled {x.shape[0]} new points")
